requirement.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.1
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
